Extension of the help and the advisor, + minor fixes.
- Extension of the advisor, to be able to call it on demand, this in order to display help on the element under the cursor.
  - On the Galaxy map, including some overlay windows.
  - On the colony panel.
  - Press "F1" to toggle. (Former help panel still available with "Ctrl-F1")
  - Options for Avatar size, text size, and to restore the former help panels.
  - This is a work in progress.
- Upgraded help panels to be able to display HTML text.
- Clicking on the Empire name in the system panel will now take you to the Diplomacy panel and select the related empire. (It was already supposed to do this, but it wasn't fully functional!)

+++ src/rotp/model/planet/PlanetFactory.java
@@ -75,9 +75,9 @@ public class PlanetFactory implements Base {
         // also change bound range (size_min, size_max) with bonus value
 		// give all sizes in multiples of inc (default = 5)
 		int inc = creationSizeInc();
-		float size_min = (float) (Math.round((creationSizeMin()*bonus)/inc) * inc);
-		float size_max = (float) (Math.round((creationSizeMax()*bonus)/inc) * inc);
-		size = (float) (Math.round((size*bonus)/inc) * inc);
+		float size_min = Math.round((creationSizeMin()*bonus)/inc) * inc;
+		float size_max = Math.round((creationSizeMax()*bonus)/inc) * inc;
+		size = Math.round((size*bonus)/inc) * inc;
 		size = Math.max(size_min, Math.min(size, size_max));
 
         p.baseSize(size);
