+++ src/rotp/model/tech/TechRepulsor.java
@@ -119,8 +119,8 @@ public final class TechRepulsor extends Tech {
 		playAudioClip(soundEffect());
 
         // calculate proper image rotation for attack frames
-        int rX = (int) (ui.boxW*3/2);
-        int rY = (int) (ui.boxH*3/2);
+        int rX = ui.boxW*3/2;
+        int rY = ui.boxH*3/2;
         float radians = source.rotateRadians(target);
 
         // create attack frames
